Update Matched V 0.3.7

+++ app/src/main/java/net/iGap/activities/ActivityManageSpace.java
@@ -9,11 +9,11 @@ import android.view.ViewGroup;
 import android.widget.CheckBox;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
 import com.afollestad.materialdialogs.DialogAction;
 import com.afollestad.materialdialogs.GravityEnum;
 import com.afollestad.materialdialogs.MaterialDialog;
-import io.realm.Realm;
-import java.io.File;
+
 import net.iGap.G;
 import net.iGap.R;
 import net.iGap.fragments.FragmentiGapMap;
@@ -21,14 +21,19 @@ import net.iGap.libs.rippleeffect.RippleView;
 import net.iGap.module.FileUtils;
 import net.iGap.module.SHP_SETTING;
 import net.iGap.realm.RealmRoomMessage;
+
 import org.osmdroid.config.Configuration;
 import org.osmdroid.config.IConfigurationProvider;
+
+import java.io.File;
+
+import io.realm.Realm;
 
 import static net.iGap.module.FileUtils.getFolderSize;
 
 public class ActivityManageSpace extends ActivityEnhanced {
     private SharedPreferences sharedPreferences;
-    private boolean isForever;
+    private int isForever;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -50,9 +55,11 @@ public class ActivityManageSpace extends ActivityEnhanced {
 
         final TextView txtSubKeepMedia = (TextView) findViewById(R.id.st_txt_sub_keepMedia);
         ViewGroup ltKeepMedia = (ViewGroup) findViewById(R.id.st_layout_keepMedia);
-        isForever = sharedPreferences.getBoolean(SHP_SETTING.KEY_KEEP_MEDIA, true);
-        if (!isForever) {
-            txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_1week));
+        isForever = sharedPreferences.getInt(SHP_SETTING.KEY_KEEP_MEDIA_NEW, 0);
+        if (isForever == 30) {
+            txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_1month));
+        } else if (isForever == 180) {
+            txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_6month));
         } else {
             txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_forever));
         }
@@ -60,11 +67,13 @@ public class ActivityManageSpace extends ActivityEnhanced {
         ltKeepMedia.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                isForever = sharedPreferences.getBoolean(SHP_SETTING.KEY_KEEP_MEDIA, true);
+                isForever = sharedPreferences.getInt(SHP_SETTING.KEY_KEEP_MEDIA_NEW, 0);
                 final int position;
-                if (!isForever) {
-                    txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_1week));
+                if (isForever == 30) {
+                    txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_1month));
                     position = 1;
+                } else if (isForever == 180) {
+                    position = 2;
                 } else {
                     txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_forever));
                     position = 0;
@@ -77,16 +86,23 @@ public class ActivityManageSpace extends ActivityEnhanced {
                         switch (which) {
                             case 0: {
                                 SharedPreferences.Editor editor = sharedPreferences.edit();
-                                editor.putBoolean(SHP_SETTING.KEY_KEEP_MEDIA, true);
+                                editor.putInt(SHP_SETTING.KEY_KEEP_MEDIA_NEW, 0);
                                 editor.apply();
                                 txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_forever));
                                 break;
                             }
                             case 1: {
                                 SharedPreferences.Editor editor = sharedPreferences.edit();
-                                editor.putBoolean(SHP_SETTING.KEY_KEEP_MEDIA, false);
+                                editor.putInt(SHP_SETTING.KEY_KEEP_MEDIA_NEW, 30);
                                 editor.apply();
-                                txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_1week));
+                                txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_1month));
+                                break;
+                            }
+                            case 2: {
+                                SharedPreferences.Editor editor = sharedPreferences.edit();
+                                editor.putInt(SHP_SETTING.KEY_KEEP_MEDIA_NEW, 180);
+                                editor.apply();
+                                txtSubKeepMedia.setText(getResources().getString(R.string.keep_media_6month));
                                 break;
                             }
                         }
@@ -224,23 +240,31 @@ public class ActivityManageSpace extends ActivityEnhanced {
                     public void onClick(View view) {
 
                         if (checkBoxPhoto.isChecked()) {
-                            for (File file : filePhoto.listFiles()) {
-                                if (!file.isDirectory()) file.delete();
+                            if (filePhoto.listFiles() != null) {
+                                for (File file : filePhoto.listFiles()) {
+                                    if (!file.isDirectory()) file.delete();
+                                }
                             }
                         }
                         if (checkBoxVideo.isChecked()) {
-                            for (File file : fileVideo.listFiles()) {
-                                if (!file.isDirectory()) file.delete();
+                            if (fileVideo.listFiles() != null) {
+                                for (File file : fileVideo.listFiles()) {
+                                    if (!file.isDirectory()) file.delete();
+                                }
                             }
                         }
                         if (checkBoxDocument.isChecked()) {
-                            for (File file : fileDocument.listFiles()) {
-                                if (!file.isDirectory()) file.delete();
+                            if (fileDocument.listFiles() != null) {
+                                for (File file : fileDocument.listFiles()) {
+                                    if (!file.isDirectory()) file.delete();
+                                }
                             }
                         }
                         if (checkBoxAudio.isChecked()) {
-                            for (File file : fileAudio.listFiles()) {
-                                if (!file.isDirectory()) file.delete();
+                            if (fileAudio.listFiles() != null) {
+                                for (File file : fileAudio.listFiles()) {
+                                    if (!file.isDirectory()) file.delete();
+                                }
                             }
                         }
                         if (checkBoxMap.isChecked()) {
